# Go: 23% speedup