--- a/Part 2/go_solution/order.go
+++ b/Part 2/go_solution/order.go
@@ -2,9 +2,6 @@ package main
 
 import (
 	"fmt"
-	"strconv"
-
-	"github.com/pkg/errors"
 )
 
 // Gary would use a fixed point type for prices in a real implementation
@@ -19,40 +16,6 @@ type Order struct {
 	Price       Price
 	IsBuy       bool
 	gen         uint64
-}
-
-func ScanOrder(r []string) (*Order, error) {
-	type Token int
-	const (
-		participantField Token = iota
-		instrumentField
-		quantityField
-		priceField
-		requiredFields
-	)
-	qty, err := strconv.Atoi(r[quantityField])
-	if err != nil {
-		return nil, errors.Wrap(err, "parsing quantity")
-	}
-	px, err := strconv.ParseFloat(r[priceField], 64)
-	if err != nil {
-		return nil, errors.Wrap(err, "parsing price")
-	}
-
-	buy := qty > 0
-	if !buy {
-		qty = -qty
-	}
-
-	return &Order{
-		IsBuy:       buy,
-		Participant: r[participantField],
-		Instrument:  r[instrumentField],
-		Quantity:    Quantity(qty),
-		Remaining:   Quantity(qty),
-		Price:       Price(px),
-	}, nil
-
 }
 
 func (o *Order) Fill(qty Quantity) error {