PR contains 2 refactoring commits and a specialised CSV reader.

Avoiding the guard rails and error handling of csv.reader gives a further speedup of 23% at the expense of 29% sanity.

This version:

```
% go build && for i in $(seq 5) ; gzip -dc ../../orders-10M.txt.gz | time ./Exchange >/dev/null
./Exchange > /dev/null  9.33s user 0.63s system 159% cpu 6.231 total
./Exchange > /dev/null  9.18s user 0.63s system 160% cpu 6.115 total
./Exchange > /dev/null  9.16s user 0.61s system 160% cpu 6.084 total
./Exchange > /dev/null  9.30s user 0.59s system 161% cpu 6.144 total
./Exchange > /dev/null  9.15s user 0.61s system 159% cpu 6.138 total
```

Previous version:

```
% go build && for i in $(seq 5) ; gzip -dc ../../orders-10M.txt.gz | time ./Exchange >/dev/null
./Exchange > /dev/null  13.42s user 0.67s system 176% cpu 7.966 total
./Exchange > /dev/null  14.31s user 0.71s system 188% cpu 7.985 total
./Exchange > /dev/null  14.79s user 0.74s system 192% cpu 8.087 total
./Exchange > /dev/null  14.39s user 0.70s system 190% cpu 7.905 total
./Exchange > /dev/null  13.86s user 0.69s system 183% cpu 7.946 total
```