@mrc 2019-11-30
The [Go CSV reader](https://golang.org/src/encoding/csv/reader.go) is pretty good at avoiding allocations but still does a bunch of work that isn't necessary given the input for trades.txt is always the same. I figure this is morally equivalent to your sscanf stuff in C++ :-)